Fix truncation bug in the new hash indexing
The bug caused progress to freeze at 63.4% when working on term 16 from term 15.

Bug introduced in c3f96c8902d50087a516fe5d831a74bf524fdead

+++ q214203-gluing-tetrahedra-together.cpp
@@ -712,7 +712,7 @@ int main(int argc, char *argv[])
                             entry = (uint8_t*)polytetTable + newPolytetCount * polytetTableElementSize;
                             break; // no duplicate of runningLeastPolytet was found in hash table
                         }
-                        entry = (uint8_t*)polytetTable + (*index - 1) * polytetTableElementSize;
+                        entry = (uint8_t*)polytetTable + (size_t)(*index - 1) * polytetTableElementSize;
                         if (memcmp(entry, runningLeastPolytet.data(), newPolytetsCompressedSize) == 0)
                             goto skipDuplicate;
                         index = (HashIndex*)((TetIndexFace*)entry + newPolytetsCompressedSize);
